snowflake-cdc-pinecone-rag/datagen/requirements.txt
Faker==25.1.0
python-dotenv
openai
snowflake-connector-python
